Funding task changes - spreadsheet link, double to decimal and validation text

+++ src/Dfe.ManageSchoolImprovement.Application/SupportProject/Commands/FundingHistory/AddFundingHistory.cs
@@ -9,7 +9,7 @@ public class AddFundingHistory
     public record AddFundingHistoryCommand(
             SupportProjectId SupportProjectId,
             string FundingType,
-            double FundingAmount,
+            decimal FundingAmount,
             string FinancialYear,
             int FundingRounds,
             string Comments
