+++ src/Dfe.ManageSchoolImprovement/Models/SupportProject/FundingHistoryViewModel.cs
@@ -6,7 +6,7 @@
         public int SupportProjectId { get; set; }
         public int ReadableId { get; set; }
         public string FundingType { get; set; }
-        public double FundingAmount { get; set; }
+        public decimal FundingAmount { get; set; }
         public string FinancialYear { get; set; }
         public int FundingRounds { get; set; }
         public string Comments { get; set; }
