+++ src/Tests/Dfe.ManageSchoolImprovement.Application.Tests/CommandHandlers/SupportProject/FundingHistory/EditFundingHistoryTests.cs
@@ -40,7 +40,7 @@ namespace Dfe.ManageSchoolImprovement.Application.Tests.SupportProject.Commands.
         {
             // Arrange
             var fundingType = "funding type";
-            var fundingAmount = 100.10;
+            var fundingAmount = (decimal)100.10;
             var financialYear = "financial year";
             var fundingRounds = 10;
             var comments = "comments";
@@ -87,7 +87,7 @@ namespace Dfe.ManageSchoolImprovement.Application.Tests.SupportProject.Commands.
         {
             // Arrange
             var fundingType = "funding type";
-            var fundingAmount = 100.10;
+            var fundingAmount = (decimal)100.10;
             var financialYear = "financial year";
             var fundingRounds = 10;
             var comments = "comments";
